calculate time punches, working on the html

+++ src/main/kotlin/com/example/shift/model/LabourSettings.kt
@@ -3,9 +3,9 @@ package com.example.shift.model
 data class LabourSettings(
         var autoBreak: Boolean,
         var dailyOvertimeMultiplier: Double,
-        var dailyOvertimeThreshold: Int,
+        var dailyOvertimeThreshold: Long,
         var overtime: Boolean,
         var weeklyOvertimeMultiplier: Double,
-        var weeklyOvertimeThreshold: Int,
+        var weeklyOvertimeThreshold: Long,
         var autoBreakRules: MutableCollection<AutoBreakRule>
 )
